chore: update dependencies
