+++ backend/package.json
@@ -19,30 +19,30 @@
   "license": "ISC",
   "description": "",
   "dependencies": {
-    "express": "^5.1.0",
+    "express": "^5.2.1",
     "morgan": "^1.10.1",
-    "pg": "^8.16.3",
+    "pg": "^8.20.0",
     "shared": "file:../shared"
   },
   "devDependencies": {
-    "@esbuild/linux-x64": "^0.25.12",
-    "@rollup/rollup-linux-x64-gnu": "^4.52.5",
-    "@types/express": "^5.0.5",
+    "@esbuild/linux-x64": "^0.27.5",
+    "@rollup/rollup-linux-x64-gnu": "^4.60.1",
+    "@types/express": "^5.0.6",
     "@types/morgan": "^1.9.10",
-    "@types/pg": "^8.15.6",
-    "@types/supertest": "^6.0.3",
-    "@typescript-eslint/eslint-plugin": "^8.46.3",
-    "@typescript-eslint/parser": "^8.46.3",
-    "@vitest/coverage-v8": "^4.0.6",
-    "@vitest/ui": "^4.0.6",
-    "eslint": "^9.39.1",
-    "globals": "^16.5.0",
-    "nodemon": "^3.1.10",
-    "supertest": "^7.1.4",
+    "@types/pg": "^8.20.0",
+    "@types/supertest": "^7.2.0",
+    "@typescript-eslint/eslint-plugin": "^8.58.0",
+    "@typescript-eslint/parser": "^8.58.0",
+    "@vitest/coverage-v8": "^4.1.2",
+    "@vitest/ui": "^4.1.2",
+    "eslint": "^10.1.0",
+    "globals": "^17.4.0",
+    "nodemon": "^3.1.14",
+    "supertest": "^7.2.2",
     "ts-node": "^10.9.2",
     "tsconfig-paths": "^4.2.0",
-    "typescript-eslint": "^8.46.3",
-    "vite-tsconfig-paths": "^5.1.4",
-    "vitest": "^4.0.6"
+    "typescript-eslint": "^8.58.0",
+    "vite-tsconfig-paths": "^6.1.1",
+    "vitest": "^4.1.2"
   }
 }
