+++ frontend/package.json
@@ -17,54 +17,54 @@
     "lint-staged": "lint-staged"
   },
   "dependencies": {
-    "@atlaskit/pragmatic-drag-and-drop": "^1.7.7",
-    "@tanstack/react-query": "^5.90.6",
+    "@atlaskit/pragmatic-drag-and-drop": "^1.7.10",
+    "@tanstack/react-query": "^5.96.1",
     "clsx": "^2.1.1",
-    "i18next": "^25.6.0",
-    "i18next-browser-languagedetector": "^8.2.0",
-    "react": "^19.2.0",
-    "react-dom": "^19.2.0",
-    "react-i18next": "^16.2.4",
-    "react-router": "^7.9.5",
+    "i18next": "^26.0.3",
+    "i18next-browser-languagedetector": "^8.2.1",
+    "react": "^19.2.4",
+    "react-dom": "^19.2.4",
+    "react-i18next": "^17.0.2",
+    "react-router": "^7.13.2",
     "react-toastify": "^11.0.5",
-    "react-virtuoso": "^4.14.1",
+    "react-virtuoso": "^4.18.4",
     "shared": "file:../shared",
     "usehooks-ts": "^3.1.1"
   },
   "devDependencies": {
-    "@commitlint/cli": "^20.1.0",
-    "@commitlint/config-conventional": "^20.0.0",
-    "@tailwindcss/vite": "^4.1.16",
-    "@tanstack/eslint-plugin-query": "^5.91.2",
+    "@commitlint/cli": "^20.5.0",
+    "@commitlint/config-conventional": "^20.5.0",
+    "@tailwindcss/vite": "^4.2.2",
+    "@tanstack/eslint-plugin-query": "^5.96.1",
     "@testing-library/dom": "^10.4.1",
     "@testing-library/jest-dom": "^6.9.1",
-    "@testing-library/react": "^16.3.0",
+    "@testing-library/react": "^16.3.2",
     "@testing-library/user-event": "^14.6.1",
-    "@types/react": "^19.2.2",
-    "@types/react-dom": "^19.2.2",
-    "@typescript-eslint/eslint-plugin": "^8.46.3",
-    "@typescript-eslint/parser": "^8.46.3",
-    "@vitejs/plugin-react": "^5.1.0",
-    "@vitest/coverage-istanbul": "^4.0.6",
-    "eslint-plugin-prettier": "^5.5.4",
+    "@types/react": "^19.2.14",
+    "@types/react-dom": "^19.2.3",
+    "@typescript-eslint/eslint-plugin": "^8.58.0",
+    "@typescript-eslint/parser": "^8.58.0",
+    "@vitejs/plugin-react": "^6.0.1",
+    "@vitest/coverage-istanbul": "^4.1.2",
+    "eslint-plugin-prettier": "^5.5.5",
     "eslint-plugin-react-hooks": "7.0.1",
-    "eslint-plugin-react-hooks-extra": "^2.3.1",
-    "eslint-plugin-react-refresh": "^0.4.24",
-    "globals": "^16.5.0",
-    "jsdom": "^27.1.0",
-    "msw": "^2.11.6",
-    "prettier": "^3.6.2",
-    "prettier-plugin-tailwindcss": "^0.7.1",
+    "eslint-plugin-react-hooks-extra": "^2.13.0",
+    "eslint-plugin-react-refresh": "^0.5.2",
+    "globals": "^17.4.0",
+    "jsdom": "^29.0.1",
+    "msw": "^2.12.14",
+    "prettier": "^3.8.1",
+    "prettier-plugin-tailwindcss": "^0.7.2",
     "steiger": "^0.5.11",
-    "tailwindcss": "^4.1.16",
-    "typescript-eslint": "^8.46.3",
-    "vite": "^7.1.12",
-    "vite-plugin-svgr": "^4.5.0",
-    "vitest": "^4.0.6"
+    "tailwindcss": "^4.2.2",
+    "typescript-eslint": "^8.58.0",
+    "vite": "^8.0.3",
+    "vite-plugin-svgr": "^5.0.0",
+    "vitest": "^4.1.2"
   },
   "optionalDependencies": {
-    "lightningcss-linux-arm64-musl": "^1.30.2",
-    "lightningcss-linux-x64-gnu": "^1.30.2",
-    "lightningcss-linux-x64-musl": "^1.30.2"
+    "lightningcss-linux-arm64-musl": "^1.32.0",
+    "lightningcss-linux-x64-gnu": "^1.32.0",
+    "lightningcss-linux-x64-musl": "^1.32.0"
   }
 }
