+++ package.json
@@ -17,18 +17,18 @@
     "stop:docker": "docker compose stop"
   },
   "dependencies": {
-    "zod": "^4.1.12"
+    "zod": "^4.3.6"
   },
   "devDependencies": {
-    "@eslint/js": "^9.39.1",
-    "@types/node": "^24.10.0",
-    "@vitest/coverage-istanbul": "^4.0.6",
-    "eslint": "^9.39.1",
-    "eslint-plugin-prettier": "^5.5.4",
+    "@eslint/js": "^10.0.1",
+    "@types/node": "^25.5.0",
+    "@vitest/coverage-istanbul": "^4.1.2",
+    "eslint": "^10.1.0",
+    "eslint-plugin-prettier": "^5.5.5",
     "husky": "^9.1.7",
-    "lint-staged": "^16.2.6",
+    "lint-staged": "^16.4.0",
     "npm-run-all": "^4.1.5",
-    "prettier": "3.6.2",
-    "typescript": "^5.9.3"
+    "prettier": "3.8.1",
+    "typescript": "^6.0.2"
   }
 }
