+++ shared/package.json
@@ -11,6 +11,6 @@
   "description": "",
   "peerDependencies": {
     "zod": "^4.0.5",
-    "typescript": "^5.8.3"
+    "typescript": ">=5.8.3"
   }
 }
